fixes bug 1204949 - Rearrange monitoring URLs
