+++ webapp-django/crashstats/crashstats/tests/test_views.py
@@ -2519,63 +2519,6 @@ class TestViews(BaseTestViews):
         response = self.client.get(url)
         eq_(response.status_code, 200)
 
-    @mock.patch('requests.get')
-    def test_crontabber_status_ok(self, rget):
-
-        def mocked_get(url, **options):
-            assert '/crontabber_state/' in url
-            return Response({
-                'state': {
-                    'job1': {
-                        'error_count': 0,
-                        'depends_on': []
-                    }
-                }
-            })
-
-        rget.side_effect = mocked_get
-
-        url = reverse('crashstats:crontabber_status')
-        response = self.client.get(url)
-        eq_(response.status_code, 200)
-        eq_(json.loads(response.content), {'status': 'ALLGOOD'})
-
-    @mock.patch('requests.get')
-    def test_crontabber_status_trouble(self, rget):
-
-        def mocked_get(url, **options):
-            assert '/crontabber_state/' in url
-            return Response({
-                'state': {
-                    'job1': {
-                        'error_count': 1,
-                        'depends_on': [],
-                    },
-                    'job2': {
-                        'error_count': 0,
-                        'depends_on': ['job1'],
-                    },
-                    'job3': {
-                        'error_count': 0,
-                        'depends_on': ['job2'],
-                    },
-                    'job1b': {
-                        'error_count': 0,
-                        'depends_on': [],
-                    },
-                }
-            })
-
-        rget.side_effect = mocked_get
-
-        url = reverse('crashstats:crontabber_status')
-        response = self.client.get(url)
-        eq_(response.status_code, 200)
-        data = json.loads(response.content)
-        eq_(data['status'], 'Broken')
-        eq_(data['broken'], ['job1'])
-        eq_(data['blocked'], ['job2', 'job3'])
-
     def test_your_crashes(self):
         url = reverse('crashstats:your_crashes')
