+++ webapp-django/crashstats/crashstats/urls.py
@@ -47,9 +47,6 @@ urlpatterns = patterns(
     url(r'^crontabber-state/$',
         views.crontabber_state,
         name='crontabber_state'),
-    url(r'^crontabber-state/status/$',
-        views.crontabber_status,
-        name='crontabber_status'),
     url(r'^your-crashes/$',
         views.your_crashes,
         name='your_crashes'),
