+++ webapp-django/crashstats/monitoring/templates/monitoring/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Things to monitor</title>
+  </head>
+  <body>
+    <ul>
+      <li><a href="{{ url('monitoring:crash_analysis_health') }}">Crash Analysis health</a></li>
+      <li><a href="{{ url('monitoring:crontabber_status') }}">Crontabber Status</a></li>
+    </ul>
+  </body>
+</html>
