+++ webapp-django/crashstats/monitoring/urls.py
@@ -4,7 +4,13 @@ from . import views
 
 urlpatterns = patterns(
     '',
+    url(r'^$',
+        views.index,
+        name='index'),
     url(r'^crash-analysis-health/$',
         views.crash_analysis_health,
         name='crash_analysis_health'),
+    url(r'^crontabber/$',
+        views.crontabber_status,
+        name='crontabber_status'),
 )
